+++ Changes.md
@@ -5,6 +5,16 @@ All notable changes to Alien::Xmake will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+## [v0.9.2] - 2026-09-02
+
+Trim the fat: distributed alien stores now hold only what end users need.
+
+### Changed
+
+- `Alien::Xrepo::Base::Builder::_resolve_alien` now re-exports installed packages as a shallow shared export (`xrepo export -k shared --shallow`) to avoid extra toolchain clutter (cmake, ninja, etc.) in the sharedir install. We're saving disk space. This might not be the *best* place to do it but it's the current place.
+
 ## [v0.9.1] - 2026-09-02
 
 More polish as I inch towards v1.0.0.
@@ -128,8 +138,10 @@ The docs have been greatly expanded since January but the stars of this release 
 ### Changed
 
 - It exists.
+[Unreleased]: https://github.com/sanko/Alien-Xmake/compare/v0.9.2...HEAD
 
-[Unreleased]: https://github.com/sanko/Alien-Xmake/compare/v0.9.1...HEAD
+[v0.9.2]: https://github.com/sanko/Alien-Xmake/compare/v0.9.1...v0.9.2
+
 [v0.9.1]: https://github.com/sanko/Alien-Xmake/compare/v0.9.0...v0.9.1
 [v0.9.0]: https://github.com/sanko/Alien-Xmake/compare/0.08...v0.9.0
 [0.08]: https://github.com/sanko/Alien-Xmake/compare/0.07...0.08